Finish BE for tabel Staff Type

+++ transportation/src/Admin/ServiceType.tsx
@@ -12,7 +12,6 @@ interface ServiceTypeData {
 }
 
 const ServiceType: React.FC = () => {
-  const [services, setServices] = useState<ServiceTypeData[]>([]);
   const [showModal, setShowModal] = useState(false);
   const [serviceTypes, setServiceTypes] = useState<ServiceTypeData[]>([]);
   const [serviceTypeCode, setServiceTypeCode] = useState("");
@@ -24,12 +23,23 @@ const ServiceType: React.FC = () => {
   const getCurrentItems = () => {
     const startIndex = (currentPage - 1) * itemsPerPage;
     const endIndex = startIndex + itemsPerPage;
-    return serviceTypes.slice(startIndex, endIndex);
+    return searchServices.slice(startIndex, endIndex);
   };
 
   const handlePageChange = (page: number) => {
     setCurrentPage(page);
   };
+
+  // Search Service
+  const [searchKeyword, setSearchKeyword] = useState<string>("");
+
+  const searchServices = serviceTypes.filter((service) =>
+    service.serviceTypeCode.toLowerCase().includes(searchKeyword.toLowerCase())
+  );
+
+  function handleSearchInputChange(e: ChangeEvent<HTMLInputElement>) {
+    setSearchKeyword(e.target.value);
+  }
 
   // Sort để lưu trữ dạng sắp xếp hiện tại
   const [sortType, setSortType] = useState<"asc" | "desc">("asc");
@@ -119,18 +129,25 @@ const ServiceType: React.FC = () => {
     }
   };
 
-  //Hàm xóa
+  // Hàm xóa
   const handleDelete = async (id: string) => {
     try {
-      const res = await axios.delete(
-        `http://localhost:4000/api/serviceType/delete/${id}`
-      );
-      console.log(res);
-      fetchData();
+      if (confirmDelete()) {
+        const res = await axios.delete(
+          `http://localhost:4000/api/serviceType/delete/${id}`
+        );
+        console.log(res);
+        fetchData();
+      }
     } catch (e) {
       console.error(e);
     }
   };
+
+  // Hàm xác nhận xóa
+  function confirmDelete() {
+    return window.confirm("Bạn có muốn xóa không??");
+  }
 
   //Kiểm tra mã trùng
   const checkDuplicateServiceTypeCode = (code: string): boolean => {
@@ -161,89 +178,100 @@ const ServiceType: React.FC = () => {
             <NavAdmin />
           </div>
           <div className="col-10">
-          <div className="container">
-            <h2 className="mt-3 mb-3">Danh sách loại dịch vụ</h2>
-            <div className="d-inline-flex">
-              <div>
+            <div className="container">
+              <h2 className="mt-3 mb-3">Danh sách loại dịch vụ</h2>
+              <div className="d-inline-flex">
+                <div>
+                  <button
+                    className="btn btn-primary"
+                    data-bs-toggle="modal"
+                    data-bs-target="#exampleModal"
+                  >
+                    Thêm loại dịch vụ
+                  </button>
+                </div>
+                <div className="ms-3 mt-1">
+                  <input
+                    className="border border-2 rounded-2"
+                    type="text"
+                    style={{ width: "300px" }}
+                    placeholder="Tìm kiếm..."
+                    value={searchKeyword}
+                    onChange={handleSearchInputChange}
+                  />
+                </div>
+              </div>
+              <table className="table mt-3">
+                <thead>
+                  <tr className="dataFields">
+                    <th
+                      style={{ cursor: "pointer" }}
+                      onClick={() => handleSort("serviceTypeCode")}
+                    >
+                      Mã loại dịch vụ <TbArrowsSort />
+                    </th>
+                    <th
+                      style={{ cursor: "pointer" }}
+                      onClick={() => handleSort("nameType")}
+                    >
+                      Tên loại dịch vụ <TbArrowsSort />
+                    </th>
+                    <th></th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {getCurrentItems().map((serviceType) => (
+                    <tr key={serviceType._id}>
+                      <td>{serviceType.serviceTypeCode}</td>
+                      <td>{serviceType.nameType}</td>
+                      <td>
+                        <button
+                          className="border border-0 bg-transparent"
+                          onClick={() => handleEdit(serviceType)}
+                        >
+                          <RiEditLine />
+                        </button>
+                      </td>
+                      <td>
+                        <button
+                          className="border border-0 bg-transparent"
+                          onClick={() => handleDelete(String(serviceType._id))}
+                        >
+                          <RiDeleteBin6Line />
+                        </button>
+                      </td>
+                    </tr>
+                  ))}
+                </tbody>
+              </table>
+
+              <div className="pagination mt-3 d-flex justify-content-center">
                 <button
-                  className="btn btn-primary"
-                  data-bs-toggle="modal"
-                  data-bs-target="#exampleModal"
+                  className="btn btn-light me-2 border border-1"
+                  onClick={() => handlePageChange(currentPage - 1)}
+                  disabled={currentPage === 1}
                 >
-                  Thêm loại dịch vụ
+                  <BiLeftArrow />
+                </button>
+                <button
+                  className="btn btn-light me-2 border border-1"
+                  onClick={() => handlePageChange(currentPage)}
+                  disabled
+                >
+                  {currentPage}
+                </button>
+                <button
+                  className="btn btn-light border border-1"
+                  onClick={() => handlePageChange(currentPage + 1)}
+                  disabled={
+                    currentPage ===
+                    Math.ceil(serviceTypes.length / itemsPerPage)
+                  }
+                >
+                  <BiRightArrow />
                 </button>
               </div>
             </div>
-            <table className="table mt-3">
-              <thead>
-                <tr className="dataFields">
-                  <th
-                    style={{ cursor: "pointer" }}
-                    onClick={() => handleSort("serviceTypeCode")}
-                  >
-                    Mã loại dịch vụ <TbArrowsSort />
-                  </th>
-                  <th
-                    style={{ cursor: "pointer" }}
-                    onClick={() => handleSort("nameType")}
-                  >
-                    Tên loại dịch vụ <TbArrowsSort />
-                  </th>
-                  <th></th>
-                </tr>
-              </thead>
-              <tbody>
-                {getCurrentItems().map((serviceType) => (
-                  <tr key={serviceType._id}>
-                    <td>{serviceType.serviceTypeCode}</td>
-                    <td>{serviceType.nameType}</td>
-                    <td>
-                      <button
-                        className="border border-0 bg-transparent"
-                        onClick={() => handleEdit(serviceType)}
-                      >
-                        <RiEditLine />
-                      </button>
-                    </td>
-                    <td>
-                      <button
-                        className="border border-0 bg-transparent"
-                        onClick={() => handleDelete(String(serviceType._id))}
-                      >
-                        <RiDeleteBin6Line />
-                      </button>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-
-            <div className="pagination mt-3 d-flex justify-content-center">
-              <button
-                className="btn btn-light me-2 border border-1"
-                onClick={() => handlePageChange(currentPage - 1)}
-                disabled={currentPage === 1}
-              >
-                <BiLeftArrow />
-              </button>
-              <button
-                className="btn btn-light me-2 border border-1"
-                onClick={() => handlePageChange(currentPage)}
-                disabled
-              >
-                {currentPage}
-              </button>
-              <button
-                className="btn btn-light border border-1"
-                onClick={() => handlePageChange(currentPage + 1)}
-                disabled={
-                  currentPage === Math.ceil(serviceTypes.length / itemsPerPage)
-                }
-              >
-                <BiRightArrow />
-              </button>
-            </div>
-          </div>
           </div>
         </div>
       </div>
